upgrade code and dependency to 2.2.10 due to incompatible API changes in PNMLFW

+++ pnml/fr.lip6.move.gal.pnml.togal/src/fr/lip6/move/gal/pnml/togal/HLGALTransformer.java
@@ -582,7 +582,7 @@ public class HLGALTransformer {
 						convertToInt(io.getSubterm().get(1), varMap));
 		} else if (g instanceof NumberConstant) {
 			NumberConstant nc = (NumberConstant) g;
-			return GF2.constant(nc.getValue());
+			return GF2.constant(nc.getValue().intValue());
 		} else if (g instanceof UserOperator) {
 			UserOperator uo = (UserOperator) g;
 			return GF2.constant(HLUtils.getConstantIndex(uo));
@@ -1041,7 +1041,7 @@ public class HLGALTransformer {
 			return fe.getElements().size();
 		} else if (psort instanceof FiniteIntRange) {
 			FiniteIntRange fir = (FiniteIntRange) psort;
-			return fir.getEnd() - fir.getStart();
+			return fir.getEnd().intValue() - fir.getStart().intValue();
 		} else if (psort instanceof ProductSort) {
 			ProductSort ps = (ProductSort) psort;
 			int sum = 1;
